+++ docs/reference/changelog.rst
@@ -33,6 +33,8 @@ Bug fixes:
 
 - The ``--recurse-submodules`` link in the ``source-submodules`` docstring
   no longer produces a linkcheck error.
+- ``CRAFT_DEFAULT_*`` environment variables are set when the default partition 
+  is aliased.
 
 For a complete list of commits, check out the `2.19.0`_ release on GitHub.